+++ munch/Week8/Week8_Mission/src/apis/lp.ts
@@ -1,83 +1,83 @@
 import type { PaginationDto } from "../types/common";
 import type {
-  RequestCreateLpDto,
-  RequestUpdateLpDto,
-  ResponseCreateLpDto,
-  ResponseLpDetailDto,
-  ResponseLpListDto,
-  ResponseUpdateLpDto,
+	RequestCreateLpDto,
+	RequestUpdateLpDto,
+	ResponseCreateLpDto,
+	ResponseLpDetailDto,
+	ResponseLpListDto,
+	ResponseUpdateLpDto,
 } from "../types/lp";
 import { axiosInstance } from "./axios";
 
 export const getLpList = async (
-  paginationDto: PaginationDto,
+	paginationDto: PaginationDto,
 ): Promise<ResponseLpListDto> => {
-  const { data } = await axiosInstance.get("/v1/lps", {
-    params: paginationDto,
-  });
-  return data;
+	const { data } = await axiosInstance.get("/v1/lps", {
+		params: paginationDto,
+	});
+	return data;
 };
 
 export const getLpDetail = async (
-  lpId: number,
+	lpId: number,
 ): Promise<ResponseLpDetailDto> => {
-  const { data } = await axiosInstance.get(`/v1/lps/${lpId}`);
-  return data;
+	const { data } = await axiosInstance.get(`/v1/lps/${lpId}`);
+	return data;
 };
 
 export const createLp = async (
-  body: RequestCreateLpDto,
+	body: RequestCreateLpDto,
 ): Promise<ResponseCreateLpDto> => {
-  const { data } = await axiosInstance.post("/v1/lps", body);
-  return data;
+	const { data } = await axiosInstance.post("/v1/lps", body);
+	return data;
 };
 
 export const updateLp = async (
-  lpId: number,
-  body: RequestUpdateLpDto,
+	lpId: number,
+	body: RequestUpdateLpDto,
 ): Promise<ResponseUpdateLpDto> => {
-  const { data } = await axiosInstance.patch(`/v1/lps/${lpId}`, body);
-  return data;
+	const { data } = await axiosInstance.patch(`/v1/lps/${lpId}`, body);
+	return data;
 };
 
 export const deleteLp = async (lpId: number): Promise<void> => {
-  await axiosInstance.delete(`/v1/lps/${lpId}`);
+	await axiosInstance.delete(`/v1/lps/${lpId}`);
 };
 
 export const getMyLpList = async (
-  paginationDto: PaginationDto,
+	paginationDto: PaginationDto,
 ): Promise<ResponseLpListDto> => {
-  const { data } = await axiosInstance.get("/v1/lps/user", {
-    params: paginationDto,
-  });
-  return data;
+	const { data } = await axiosInstance.get("/v1/lps/user", {
+		params: paginationDto,
+	});
+	return data;
 };
 
 export const likeLp = async (lpId: number) => {
-  const { data } = await axiosInstance.post(`/v1/lps/${lpId}/likes`);
-  return data;
+	const { data } = await axiosInstance.post(`/v1/lps/${lpId}/likes`);
+	return data;
 };
 
 export const unlikeLp = async (lpId: number) => {
-  const { data } = await axiosInstance.delete(`/v1/lps/${lpId}/likes`);
-  return data;
+	const { data } = await axiosInstance.delete(`/v1/lps/${lpId}/likes`);
+	return data;
 };
 
 export const getMyLikedLpList = async (
-  paginationDto: PaginationDto,
+	paginationDto: PaginationDto,
 ): Promise<ResponseLpListDto> => {
-  const { data } = await axiosInstance.get("/v1/lps/likes/me", {
-    params: paginationDto,
-  });
-  return data;
+	const { data } = await axiosInstance.get("/v1/lps/likes/me", {
+		params: paginationDto,
+	});
+	return data;
 };
 
 export const getLpListByTag = async (
-  tagName: string,
-  paginationDto: PaginationDto,
+	tagName: string,
+	paginationDto: PaginationDto,
 ): Promise<ResponseLpListDto> => {
-  const { data } = await axiosInstance.get(`/v1/lps/tag/${tagName}`, {
-    params: paginationDto,
-  });
-  return data;
+	const { data } = await axiosInstance.get(`/v1/lps/tag/${tagName}`, {
+		params: paginationDto,
+	});
+	return data;
 };
