+++ munch/Week8/Week8_Mission/src/hooks/useDebounce.ts
@@ -1,17 +1,17 @@
 import { useEffect, useState } from "react";
 
 function useDebounce<T>(value: T, delay: number = 300): T {
-  const [debounceValue, setDebounceValue] = useState<T>(value);
+	const [debounceValue, setDebounceValue] = useState<T>(value);
 
-  useEffect(() => {
-    const timer = setTimeout(() => {
-      setDebounceValue(value);
-    }, delay);
+	useEffect(() => {
+		const timer = setTimeout(() => {
+			setDebounceValue(value);
+		}, delay);
 
-    return () => clearTimeout(timer);
-  }, [value, delay]);
+		return () => clearTimeout(timer);
+	}, [value, delay]);
 
-  return debounceValue;
+	return debounceValue;
 }
 
 export default useDebounce;
