+++ munch/Week8/Week8_Mission/src/hooks/useIntersectionObserver.ts
@@ -1,31 +1,31 @@
 import { useEffect, useRef } from "react";
-import useThrottle from "./useThrotle";
+import useThrottle from "./useThrottle";
 
 export const useIntersectionObserver = (
-  callback: () => void,
-  hasNextPage: boolean | undefined,
-  interval: number = 1000,
+	callback: () => void,
+	hasNextPage: boolean | undefined,
+	interval: number = 1000,
 ) => {
-  const observerRef = useRef<HTMLDivElement>(null);
+	const observerRef = useRef<HTMLDivElement>(null);
 
-  const throttledCallback = useThrottle(callback, interval);
+	const throttledCallback = useThrottle(callback, interval);
 
-  useEffect(() => {
-    const observer = new IntersectionObserver(
-      (entries) => {
-        if (entries[0].isIntersecting && hasNextPage) {
-          throttledCallback();
-        }
-      },
-      { threshold: 0.5 },
-    );
+	useEffect(() => {
+		const observer = new IntersectionObserver(
+			(entries) => {
+				if (entries[0].isIntersecting && hasNextPage) {
+					throttledCallback();
+				}
+			},
+			{ threshold: 0.5 },
+		);
 
-    if (observerRef.current) {
-      observer.observe(observerRef.current);
-    }
+		if (observerRef.current) {
+			observer.observe(observerRef.current);
+		}
 
-    return () => observer.disconnect();
-  }, [throttledCallback, hasNextPage]);
+		return () => observer.disconnect();
+	}, [throttledCallback, hasNextPage]);
 
-  return observerRef;
+	return observerRef;
 };
